Backend_with_scrapper_n_LLM/requirements.txt
flask
flask-socketio
python-dotenv
playwright
langchain
langchain-groq 